requirements.txt
# Core packages
pandas
tqdm

# OCR and image handling
pytesseract
Pillow
pdf2image

# Matching engine
rapidfuzz

# Optional: for web UI
streamlit
